requirements.txt
asn1crypto==0.23.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.3
enum34==1.1.6
future==0.16.0
idna==2.6
ipaddress==1.0.18
oauthlib==2.0.6
pycparser==2.18
pyOpenSSL==17.4.0
pysolr==3.6.0
python-dateutil==2.6.1
python-twitter==3.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
watson-developer-cloud==1.0.0
